refactor: rearrange contracts and imporve exports
