+++ typescript/optics-provider/src/optics/contracts/BridgeContracts.ts
@@ -5,8 +5,8 @@ import {
   BridgeRouter__factory,
   ETHHelper,
   ETHHelper__factory,
-} from '../../../typechain/optics-xapps';
-import { Contracts } from '../contracts/contracts';
+} from '../../../../typechain/optics-xapps';
+import { Contracts } from '../../contracts';
 
 type Address = string;
