+++ typescript/optics-provider/src/optics/index.ts
@@ -1,4 +1,8 @@
-export { BridgeContracts } from './BridgeContracts';
-export { CoreContracts } from './CoreContracts';
+export { BridgeContracts } from './contracts/BridgeContracts';
+export { CoreContracts } from './contracts/CoreContracts';
+export { BridgeMessage } from './messages/BridgeMessage';
+export { OpticsMessage } from './messages/message';
+export { ResolvedTokenInfo, TokenIdentifier } from './tokens';
+
 export { OpticsDomain, mainnetDomains } from './OpticsDomain';
 export { OpticsContext, mainnet } from './OpticsContext';
