+++ typescript/optics-provider/src/optics/tokens.ts
@@ -1,5 +1,5 @@
 import { BytesLike } from 'ethers';
-import { ERC20 } from '../../typechain/optics-xapps';
+import { ERC20 } from '../../../typechain/optics-xapps';
 
 export interface TokenIdentifier {
   domain: string | number;
